backend/requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
numpy==1.26.0
opencv-python==4.9.0.80
insightface==0.7.3
onnxruntime-gpu>=1.17.0
websockets==12.0
pydantic==2.6.3
python-jose==3.3.0
httpx==0.27.0
ffmpeg-python==0.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
mangum==0.17.0
slowapi==0.1.8
passlib==1.7.4
Jinja2==3.1.3
aiofiles==23.2.1
